% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moc_price.R
\name{get_moc_price}
\alias{get_moc_price}
\title{get_moc_price}
\usage{
get_moc_price(
  product_id,
  from_date = "2021-01-01",
  to_date = Sys.Date(),
  meta = FALSE
)
}
\arguments{
\item{product_id}{a product id}

\item{from_date}{starting date , Default: '2021-01-01'}

\item{to_date}{end date, Default: Sys.Date()}

\item{meta}{a , Default: FALSE}
}
\value{
a price data in data.frame
}
\description{
get a price data from moc opendata
}
\examples{
\dontrun{
if(interactive()) {
# get white rice 100\% grad A (new rice) 
 price_R11002 <- get_moc_price("R11002")
 }
}
}
